+++ saitama_neko/CreateSentens.py
@@ -1,0 +1,33 @@
+import SaitamaSeCats
+
+
+# 文章作成クラス
+class CreateSentens(SaitamaSeCats.SaitamaCats):
+    def __init__(self,area):
+        super().__init__(area)
+        self.SENT_1 = '現在、埼玉県の{}地区では{}匹の猫が飼い主を募集しています'
+        self.SENT_2 = '【今日の埼玉県{}地区の譲渡用猫情報】\n募集中          : {} 匹\nお見合い中      : {} 匹\n飼い主さん決定  : {} 匹\n{}'
+        self.area_jp = self.transrate()
+    
+    
+    def transrate(self):
+        if self.area == 'nw':
+            jp = '北部・西部'
+        elif self.area == 'se':
+            jp = '南部・東部'
+        return jp
+
+    
+    # 文章作成1
+    def sentens_1(self):
+        super().count_cats()
+        sent = self.SENT_1.format(self.area_jp,self.cats_count_all)
+        return sent
+
+
+    # 文章作成2
+    def sentens_2(self):
+        for filter in ['wanted', 'interview', 'decided']:
+            super().count_cats(filter)
+        sent = self.SENT_2.format(self.area_jp,self.cats_count_wanted,self.cats_count_interview,self.cats_count_decided,self.url)
+        return sent
